Removed stripe secret key

+++ DemoAPI/ProductService/Models/Product.cs
@@ -7,8 +7,9 @@ namespace ProductService.Models
 
         public string Name { get; set; }
         
-        public int Price { get; set; }
+        public decimal Price { get; set; }
 
+        
         public string PictureUrl { get; set; }
 
         public int BrandId { get; set; }
